员工打卡规则 添加排班信息。员工打卡规则实体与企业所有打卡规则实体抽取通用对象。
